% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildApricotR-package.R
\docType{package}
\name{wildApricotR-package}
\alias{wildApricotR}
\alias{wildApricotR-package}
\title{wildApricotR: Data mining on a Wild Apricot CMS installation}
\description{
Access Wild Apricot CMS data from the R programming language. https://wildapricot.com
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mnr/wildApricotR}
  \item \url{https://mnr.github.io/wildApricotR/}
  \item Report bugs at \url{https://github.com/mnr/wildApricotR/issues}
}

}
\keyword{internal}
